refactor vector_field_function
