+++ dynamo/prediction/state_graph.py
@@ -4,7 +4,7 @@ from scipy.spatial import cKDTree
 from tqdm import tqdm
 
 from dynamo.prediction.fate import _fate
-from dynamo.tools.scVectorField import vector_field_function
+from dynamo.tools import vector_field_function
 from dynamo.tools.utils import (
     fetch_states,
     remove_redundant_points_trajectory,
